mpc: move hash functions and hash test into separate files
